Imported Upstream version 3.1.0-dfsg

+++ src/VBox/Runtime/r0drv/darwin/RTLogWriteStdOut-r0drv-darwin.cpp
@@ -1,4 +1,4 @@
-/* $Id: RTLogWriteStdOut-r0drv-darwin.cpp $ */
+/* $Id: RTLogWriteStdOut-r0drv-darwin.cpp 22052 2009-08-07 09:45:48Z vboxsync $ */
 /** @file
  * IPRT - Log To StdOut, Ring-0 Driver, Darwin.
  */
@@ -33,13 +33,13 @@
 *   Header Files                                                               *
 *******************************************************************************/
 #include "the-darwin-kernel.h"
+#include "internal/iprt.h"
 #include <iprt/log.h>
-#include <iprt/assert.h>
 
 
 RTDECL(void) RTLogWriteStdOut(const char *pch, size_t cb)
 {
-    printf("%.*s", cb, pch);
+    printf("%.*s", (int)cb, pch);
     return;
 }
